fix: svg load
