+++ electron/main/log.d.ts
@@ -1,0 +1,1 @@
+export declare function Debug(str: string): void
