+++ electron/main/log.ts
@@ -1,6 +1,6 @@
 const isDevelopment = process.env.NODE_ENV !== 'production'
 
-export const Debug = (str: string) => {
+export const Debug = (str: string): void => {
   if (isDevelopment) {
     console.log(str)
   }
