+++ package.json
@@ -12,8 +12,8 @@
     "dev": "vite",
     "build": "vue-tsc --noEmit && vite build && electron-builder",
     "build:all": "vue-tsc --noEmit && vite build && electron-builder -mwl",
-    "lint": "eslint src",
-    "lint:fix": "eslint src electron --fix --ext .ts,.tsx,.js",
+    "lint": "eslint src electron  --ext .ts,.tsx,.vue",
+    "lint:fix": "eslint src electron --fix --ext .ts,.tsx,.vue",
     "release": "vue-tsc --noEmit && vite build && electron-builder --publish onTagOrDraft",
     "patch": "npm version patch && git push origin master && git push origin --tags",
     "minor": "npm version minor && git push origin master && git push origin --tags",
@@ -36,7 +36,7 @@
     "@types/sqlite3": "^3.1.8",
     "@typescript-eslint/eslint-plugin": "^5.30.6",
     "@typescript-eslint/parser": "^5.30.6",
-    "@vitejs/plugin-vue": "^3.0.0",
+    "@vitejs/plugin-vue": "^2.0.0",
     "@vue/devtools": "^6.2.1",
     "ant-design-vue": "^3.0.0",
     "axios": ">=0.21.1",
@@ -61,9 +61,10 @@
     "unplugin-vue-components": "^0.21.2",
     "uuid": "^8.0.0",
     "vditor": "^3.0.0",
-    "vite": "^3.0.0",
+    "vite": "^2.0.0",
     "vite-plugin-electron": "^0.8.0",
-    "vue": "^3.2.37",
+    "vite-plugin-svg-icons": "^2.0.1",
+    "vue": "^3.0.0",
     "vue-router": "^4.0.0",
     "vue-tsc": "^0.38.3"
   },
