+++ src/assets/css/index.styl
@@ -62,7 +62,7 @@
     --sideBarIconColor var(--iconColor)
     --sideBarTitleColor rgba(255, 255, 255, 1)
     --sideBarTextColor rgba(214, 214, 214, 1)
-    --sideBarBgColor #50616d
+    --sideBarBgColor rgba(46, 50, 53, 1)
     --sideBarItemHoverBgColor rgba(0, 0, 0, 0.03)
     --itemBgColor rgba(255, 255, 255, 0.6)
     --floatFontColor rgba(0, 0, 0, 0.6)
