+++ src/main.ts
@@ -4,12 +4,15 @@ import App from './App.vue'
 import router from '@/router'
 
 import devtools from '@vue/devtools'
+console.log(process.env.NODE_ENV)
 if (process.env.NODE_ENV === 'development') {
   devtools.connect()
 }
 
 import '@/plugins/sqlite3/init'
 import '@/assets/css/index.styl'
+
+import 'virtual:svg-icons-register'
 
 import pinia from '@/store'
