--- src/preferences/sideBar/config.js
@@ -1,10 +1,0 @@
-import GeneralIcon from '@/assets/icons/pref_general.svg'
-
-export const category = [
-  {
-    name: 'General',
-    label: 'general',
-    icon: GeneralIcon,
-    path: '/preference/general',
-  },
-]
