+++ src/preferences/sideBar/config.ts
@@ -1,0 +1,11 @@
+// import GeneralIcon from '@/assets/icons/pref_general.svg'
+// console.log(GeneralIcon)
+
+export const category = [
+  {
+    name: 'General',
+    label: 'general',
+    icon: 'pref_general',
+    path: '/preference/general',
+  },
+]
