+++ tsconfig.json
@@ -21,14 +21,18 @@
     "allowSyntheticDefaultImports": true,
     "skipLibCheck": true,
     "noImplicitAny": false,
+    "allowJs": true
   },
-  // "include": [
-  //   "src/**/*.ts",
-  //   "src/**/*.d.ts",
-  //   "src/**/*.tsx",
-  //   "src/**/*.vue",
-  //   "src/**/*.js",
-  // ],
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.d.ts",
+    "src/**/*.tsx",
+    "src/**/*.vue",
+    "src/**/*.js",
+  ],
+  "exclude": [
+    "src/bootstrap.js",
+  ],
   "references": [
     {
       "path": "./tsconfig.node.json"
